pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["chado13 <geaya13@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
bs4 = "^0.0.2"
lxml = "^5.3.0"
discord = "^2.3.2"
typer = "^0.13.0"
discord-webhook = "^1.3.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
